pyproject.toml
[project]
name = "projeto-api-tdd"
version = "0.1.0"
description = ""
authors = [
    {name = "luiz",email = "luizfilipefmg@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "motor (>=3.7.1,<4.0.0)"
]

[tool.poetry]
packages = [{include = "projeto_api_tdd", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
